pkg/embedded/templates/ml-image/pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-image-project"
version = "0.1.0"
description = "ML Image Processing Project"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "pytorch-lightning>=2.2.0",
    "opencv-python>=4.9.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    "wandb>=0.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]
